adapt to libfirm-1.19.1 release

+++ src/firm/GraphBase.java
@@ -186,7 +186,7 @@ public abstract class GraphBase extends JNAWrapper {
 	}
 
 	public int getIdx() {
-		return binding_irgraph.get_irg_idx(ptr).intValue();
+		return binding_irgraph.get_irg_idx(ptr);
 	}
 
 	public void incVisited() {
